smush all the non-reusable mixins into resource
