+++ json_api_client.gemspec
@@ -15,6 +15,7 @@ Gem::Specification.new do |s|
   s.add_dependency "faraday", '>= 0.8.0'
   s.add_dependency "faraday_middleware", '~> 0.9'
   s.add_dependency "addressable", '~> 2.2'
+  s.add_dependency "activemodel", '>= 3.2.0'
 
   s.add_development_dependency "webmock"
   s.add_development_dependency "mocha"
