+++ lib/json_api_client/helpers/dynamic_attributes.rb
@@ -7,7 +7,7 @@ module JsonApiClient
       end
 
       def attributes=(attrs = {})
-        @attributes ||= {}.with_indifferent_access
+        @attributes ||= ActiveSupport::HashWithIndifferentAccess.new
 
         return @attributes unless attrs.present?
         attrs.each do |key, value|
