--- lib/json_api_client/helpers/inspectable.rb
@@ -1,11 +1,0 @@
-module JsonApiClient
-  module Helpers
-    module Inspectable
-
-      def inspect
-        "#<#{self.class.name}:@attributes=#{attributes.inspect}>"
-      end
-
-    end
-  end
-end
